% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{session_durations}
\alias{session_durations}
\title{Brushing durations for 13 subjects, 10 sessions each}
\format{
TODO: complete this
}
\usage{
session_durations
}
\description{
Brushing durations for 13 subjects, 10 sessions each
}
\keyword{datasets}
